requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
blis==1.0.1
cachetools==5.5.0
catalogue==2.0.10
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cloudpathlib==0.20.0
colorama==0.4.6
comm==0.2.2
confection==0.1.5
cymem==2.0.10
debugpy==1.8.1
decorator==5.1.1
Django==5.1.3
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
example==0.1.0
executing==2.0.1
Flask==2.3.3
frozenlist==1.4.0
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.68.0
grpcio-status==1.68.0
httplib2==0.22.0
idna==3.4
ipykernel==6.29.4
ipython==8.23.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.6.1
jupyter_core==5.7.2
langcodes==3.5.0
language_data==1.3.0
lxml==5.3.0
magic-filter==1.0.11
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.0.4
murmurhash==1.0.11
nest-asyncio==1.6.0
numpy==2.0.2
openai==0.28.1
packaging==24.0
parso==0.8.3
platformdirs==4.2.0
preshed==3.0.9
prompt-toolkit==3.0.43
proto-plus==1.25.0
protobuf==5.29.0
psutil==5.9.8
psycopg2-binary==2.9.10
pure-eval==0.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.3.0
pydantic_core==2.6.3
Pygments==2.17.2
pyparsing==3.2.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
pywin32==306
pyzmq==25.1.2
requests==2.31.0
rich==13.9.4
rsa==4.9
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.2
srsly==2.4.8
stack-data==0.6.3
thinc==8.3.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.2
typer==0.14.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.0.5
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
Werkzeug==2.3.7
wrapt==1.17.0
yarl==1.9.2
